requirements.txt
fastapi
openai
python-dotenv==1.0.0
pydantic==2.5.0
uvicorn==0.24.0
jinja2
python-multipart==0.0.6